% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolvio.R
\docType{data}
\name{schoolvio}
\alias{schoolvio}
\title{Dataset for Shooting Incidents at K-12 Schools 1970-Present}
\format{
Various \code{tibble} listed below:
\describe{
\item{schoolvio}{Cleaned Master Dataset}
\item{svioid}{Cleaned Incident Report Dataset}
\item{svioshooter}{Cleaned Shooter Dataset}
\item{sviovictim}{Cleaned Victim Dataset}
\item{svioweapon}{Cleaned Weapon Dataset}
\item{svioraw}{Raw uncleaned Dataset}
}
}
\usage{
schoolvio
}
\description{
Dataset publishing of the Center for Homeland Defense and Security's K-12 School Shooting Database
}
\keyword{datasets}
